+++ irodsbackend/tests/test_permissions_taskflow.py
@@ -4,17 +4,13 @@ from django.test import override_settings
 
 # Projectroles dependency
 from projectroles.models import SODAR_CONSTANTS
-from projectroles.tests.test_permissions import TestPermissionMixin
 
 # Samplesheets dependency
-from samplesheets.tests.test_io import (
-    SampleSheetIOMixin,
-    SHEET_DIR,
-)
+from samplesheets.tests.test_io import SampleSheetIOMixin, SHEET_DIR
 from samplesheets.tests.test_views_taskflow import SampleSheetTaskflowMixin
 
 # Taskflowbackend dependency
-from taskflowbackend.tests.base import TaskflowbackendTestBase
+from taskflowbackend.tests.base import TaskflowPermissionTestBase
 
 
 # SODAR constants
@@ -29,49 +25,18 @@ NON_PROJECT_PATH = '/sodarZone/projects'
 
 
 class TestIrodsbackendPermissions(
-    TestPermissionMixin,
     SampleSheetIOMixin,
     SampleSheetTaskflowMixin,
-    TaskflowbackendTestBase,
+    TaskflowPermissionTestBase,
 ):
     """Tests for irodsbackend API view permissions"""
 
     def setUp(self):
         super().setUp()
-        # Init users
-        self.superuser = self.user  # HACK
-        self.anonymous = None
-        self.user_owner = self.make_user('user_owner')
-        self.user_delegate = self.make_user('user_delegate')
-        self.user_contributor = self.make_user('user_contributor')
-        self.user_guest = self.make_user('user_guest')
-        self.user_no_roles = self.make_user('user_no_roles')
-
-        # Set up project with taskflow
-        self.project, self.owner_as = self.make_project_taskflow(
-            title='TestProject',
-            type=PROJECT_TYPE_PROJECT,
-            parent=self.category,
-            owner=self.user_owner,
-            description='description',
-        )
-
-        # Set up assignments with taskflow
-        self.delegate_as = self.make_assignment_taskflow(
-            self.project, self.user_delegate, self.role_delegate
-        )
-        self.contributor_as = self.make_assignment_taskflow(
-            self.project, self.user_contributor, self.role_contributor
-        )
-        self.guest_as = self.make_assignment_taskflow(
-            self.project, self.user_guest, self.role_guest
-        )
-
         # Set up investigation
         self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
         # Create iRODS collections
         self.make_irods_colls(self.investigation)
-
         # Set up test paths
         self.project_path = self.irods_backend.get_path(self.project)
         self.sample_path = self.irods_backend.get_sample_path(self.project)
@@ -83,13 +48,16 @@ class TestIrodsbackendPermissions(
         )
         good_users = [
             self.superuser,
-            self.user_owner_cat,  # Inherited owner
+            self.user_owner_cat,  # Inherited
+            self.user_delegate_cat,  # Inherited
+            self.user_contributor_cat,  # Inherited
+            self.user_guest_cat,  # Inherited
             self.user_owner,
             self.user_delegate,
             self.user_contributor,
             self.user_guest,
         ]
-        bad_users = [self.user_no_roles, self.anonymous]
+        bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
         self.assert_response(url, good_users, 200)
         self.assert_response(url, bad_users, 403)
         self.project.set_public()
@@ -98,7 +66,7 @@ class TestIrodsbackendPermissions(
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
     def test_stats_get_anon(self):
-        """Test stats API view with anonymous access"""
+        """Test stats Ajax view with anonymous access"""
         self.project.set_public()
         url = self.irods_backend.get_url(
             view='stats', project=self.project, path=self.sample_path
@@ -107,7 +75,7 @@ class TestIrodsbackendPermissions(
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
     def test_stats_get_anon_no_perms(self):
-        """Test stats API view with anonymous access and no perms to collection"""
+        """Test stats Ajax view with anonymous access and no collection perms"""
         self.project.set_public()
         url = self.irods_backend.get_url(
             view='stats', project=self.project, path=self.project_path
@@ -115,13 +83,17 @@ class TestIrodsbackendPermissions(
         self.assert_response(url, self.anonymous, 403)
 
     def test_stats_get_not_in_project(self):
-        """Test stats API view GET with path not in project"""
+        """Test stats Ajax view GET with path not in project"""
         url = self.irods_backend.get_url(
             view='stats', project=self.project, path=NON_PROJECT_PATH
         )
         bad_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
             self.user_owner,
             self.user_delegate,
             self.user_contributor,
@@ -132,7 +104,7 @@ class TestIrodsbackendPermissions(
         self.assert_response(url, bad_users, 400)
 
     def test_stats_get_no_perms(self):
-        """Test stats API view GET without collection perms"""
+        """Test stats Ajax view GET without collection perms"""
         test_path = self.project_path + '/' + TEST_COLL_NAME
         self.irods.collections.create(test_path)  # NOTE: No perms given
 
@@ -142,10 +114,14 @@ class TestIrodsbackendPermissions(
         good_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
             self.user_owner,
             self.user_delegate,
         ]
         bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
             self.user_contributor,
             self.user_guest,
             self.user_no_roles,
@@ -155,7 +131,7 @@ class TestIrodsbackendPermissions(
         self.assert_response(url, bad_users, 403)
 
     def test_stats_post(self):
-        """Test stats API view POST"""
+        """Test stats Ajax view POST"""
         url = self.irods_backend.get_url(
             view='stats',
             project=self.project,
@@ -167,31 +143,37 @@ class TestIrodsbackendPermissions(
         good_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
             self.user_owner,
             self.user_delegate,
             self.user_contributor,
             self.user_guest,
         ]
-        bad_users = [self.user_no_roles, self.anonymous]
+        bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
         self.assert_response(
             url, good_users, 200, method='POST', data=post_data
         )
         self.assert_response(url, bad_users, 403, method='POST', data=post_data)
 
     def test_list_get(self):
-        """Test object list API view GET"""
+        """Test object list GET"""
         url = self.irods_backend.get_url(
             view='list', project=self.project, path=self.sample_path, md5=0
         )
         good_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
             self.user_owner,
             self.user_delegate,
             self.user_contributor,
             self.user_guest,
         ]
-        bad_users = [self.user_no_roles, self.anonymous]
+        bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
         self.assert_response(url, good_users, 200)
         self.assert_response(url, bad_users, 403)
         self.project.set_public()
@@ -200,7 +182,7 @@ class TestIrodsbackendPermissions(
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
     def test_list_get_anon(self):
-        """Test object list API view GET with anonymous access"""
+        """Test object list GET with anonymous access"""
         url = self.irods_backend.get_url(
             view='list', project=self.project, path=self.sample_path, md5=0
         )
@@ -209,7 +191,7 @@ class TestIrodsbackendPermissions(
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
     def test_list_get_anon_no_perms(self):
-        """Test object list API view GET with anonymous access and no permission"""
+        """Test object list GET with anon access and no permission"""
         url = self.irods_backend.get_url(
             view='list', project=self.project, path=self.project_path, md5=0
         )
@@ -224,6 +206,10 @@ class TestIrodsbackendPermissions(
         bad_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
             self.user_owner,
             self.user_delegate,
             self.user_contributor,
@@ -244,10 +230,14 @@ class TestIrodsbackendPermissions(
         good_users = [
             self.superuser,
             self.user_owner_cat,
+            self.user_delegate_cat,
             self.user_owner,
             self.user_delegate,
         ]
         bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
             self.user_contributor,
             self.user_guest,
             self.user_no_roles,