+++ taskflowbackend/api.py
@@ -2,12 +2,6 @@
 
 import json
 import logging
-import os
-
-from irods.models import TicketQuery, UserGroup
-
-from django.conf import settings
-from django.core.exceptions import ImproperlyConfigured
 
 # Landingzones dependency
 from landingzones.models import LandingZone
@@ -29,7 +23,6 @@ logger = logging.getLogger(__name__)
 PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 
 # Local constants
-DEFAULT_PERMANENT_USERS = ['client_user', 'rods', 'rodsadmin', 'public']
 UNKNOWN_RUN_ERROR = 'Running flow failed: unknown error, see server log'
 LOCK_FAIL_MSG = 'Unable to acquire project lock'
 
@@ -249,72 +242,6 @@ class TaskflowAPI:
             async_mode=False,
             tl_event=tl_event,
         )
-
-    @classmethod
-    def cleanup(cls):
-        """
-        Send a cleanup command to SODAR Taskflow. Only allowed in test mode.
-
-        :return: Boolean
-        :raise: ImproperlyConfigured if TASKFLOW_TEST_MODE is not set True
-        :raise: Exception if iRODS cleanup fails
-        """
-        if not settings.TASKFLOW_TEST_MODE:
-            raise ImproperlyConfigured(
-                'TASKFLOW_TEST_MODE not True, cleanup command not allowed'
-            )
-        irods_backend = get_backend_api('omics_irods')
-        projects_root = irods_backend.get_projects_path()
-        permanent_users = getattr(
-            settings, 'TASKFLOW_TEST_PERMANENT_USERS', DEFAULT_PERMANENT_USERS
-        )
-        # TODO: Remove stuff from user home collections
-
-        with irods_backend.get_session() as irods:
-            # Remove project folders
-            try:
-                irods.collections.remove(
-                    projects_root, recurse=True, force=True
-                )
-                logger.debug('Removed projects root: {}'.format(projects_root))
-            except Exception:
-                pass  # This is OK, the root just wasn't there
-                # Remove created user groups and users
-
-            # NOTE: user_groups.remove does both
-            for g in irods.query(UserGroup).all():
-                if g[UserGroup.name] not in permanent_users:
-                    irods.user_groups.remove(user_name=g[UserGroup.name])
-                    logger.debug('Removed user: {}'.format(g[UserGroup.name]))
-
-            # Remove all tickets
-            ticket_query = irods.query(TicketQuery.Ticket).all()
-            for ticket in ticket_query:
-                ticket_str = ticket[TicketQuery.Ticket.string]
-                irods_backend.delete_ticket(irods, ticket_str)
-                logger.debug('Deleted ticket: {}'.format(ticket_str))
-
-            # Remove data objects and unneeded collections from trash
-            trash_path = irods_backend.get_trash_path()
-            trash_coll = irods.collections.get(trash_path)
-            # NOTE: We can't delete the home trash collection
-            trash_home_path = os.path.join(trash_path, 'home')
-            for coll in irods_backend.get_colls_recursively(trash_coll):
-                if irods.collections.exists(
-                    coll.path
-                ) and not coll.path.startswith(trash_home_path):
-                    irods.collections.remove(
-                        coll.path, recurse=True, force=True
-                    )
-            obj_paths = [
-                o['path']
-                for o in irods_backend.get_objs_recursively(irods, trash_coll)
-                + irods_backend.get_objs_recursively(
-                    irods, trash_coll, md5=True
-                )
-            ]
-            for path in obj_paths:
-                irods.data_objects.unlink(path, force=True)
 
     @classmethod
     def get_error_msg(cls, flow_name, submit_info):